requirements.txt
astroid==2.3.3
isort==4.3.21
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.4.4
PyYAML==5.3.1
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
